Trunk/Banks & Merchant mode

+++ graviton - game/src/main/java/org/graviton/game/exchange/Exchange.java
@@ -15,7 +15,7 @@ public interface Exchange {
 
     void removeItem(int itemId, short quantity, int exchangerId);
 
-    void editKamas(int quantity, int exchangerId);
+    void editKamas(long quantity, int exchangerId);
 
     void toggle(int exchangerId);
